+++ kite/src/main/kotlin/com/tang/kite/wrapper/where/AbstractWhereWrapper.kt
@@ -292,6 +292,66 @@ abstract class AbstractWhereWrapper<T, R, W>(
     }
 
     /**
+     * Order by ascending with column name
+     *
+     * @param column column name
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderByAsc(column: String): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, true)
+    }
+
+    /**
+     * Order by descending with column name
+     *
+     * @param column column name
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderByDesc(column: String): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, false)
+    }
+
+    /**
+     * Order by ascending with property reference
+     *
+     * @param column column property
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun <E> orderByAsc(column: KMutableProperty1<E, *>): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, true)
+    }
+
+    /**
+     * Order by descending with property reference
+     *
+     * @param column column property
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun <E> orderByDesc(column: KMutableProperty1<E, *>): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, false)
+    }
+
+    /**
+     * Order by ascending with SFunction
+     *
+     * @param column column function
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun <E> orderByAsc(column: SFunction<E, *>): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, true)
+    }
+
+    /**
+     * Order by descending with SFunction
+     *
+     * @param column column function
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun <E> orderByDesc(column: SFunction<E, *>): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, false)
+    }
+
+    /**
      * Get the SQL statement
      *
      * @return SqlStatement