+++ kite/src/main/kotlin/com/tang/kite/wrapper/where/WhereHavingWrapper.kt
@@ -38,7 +38,7 @@ class WhereHavingWrapper<T, R, W>(
     }
 
     private fun getColumnName(column: KMutableProperty1<T, *>): String {
-        var field = column.javaField!!
+        val field = column.javaField!!
         return getColumnName(field)
     }
 
@@ -247,6 +247,66 @@ class WhereHavingWrapper<T, R, W>(
     }
 
     /**
+     * Order by ascending with column name
+     *
+     * @param column column name
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderByAsc(column: String): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, true)
+    }
+
+    /**
+     * Order by descending with column name
+     *
+     * @param column column name
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderByDesc(column: String): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, false)
+    }
+
+    /**
+     * Order by ascending with property reference
+     *
+     * @param column column property
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderByAsc(column: KMutableProperty1<T, *>): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, true)
+    }
+
+    /**
+     * Order by descending with property reference
+     *
+     * @param column column property
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderByDesc(column: KMutableProperty1<T, *>): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, false)
+    }
+
+    /**
+     * Order by ascending with SFunction
+     *
+     * @param column column function
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderByAsc(column: SFunction<T, *>): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, true)
+    }
+
+    /**
+     * Order by descending with SFunction
+     *
+     * @param column column function
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderByDesc(column: SFunction<T, *>): WhereOrderByWrapper<T, R, W> {
+        return orderBy(column, false)
+    }
+
+    /**
      * Build the wrapper
      *
      * @return W
@@ -261,7 +321,6 @@ class WhereHavingWrapper<T, R, W>(
      *
      * @return R
      */
-    @Suppress("UNCHECKED_CAST")
     override fun execute(): R {
         return whereWrapper.execute()
     }