+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 
  * Fix `Fields.getValue` not throwing an exception when the map does not contain the key
+ * Fix generic type error in `orderByAsc` and `orderByDesc` methods
 
 ### Dependencies